+++ layer-admin/src/main/java/org/layer/space/controller/AdminSpaceController.java
@@ -1,14 +1,12 @@
 package org.layer.space.controller;
 
 import lombok.RequiredArgsConstructor;
+import org.layer.retrospect.controller.dto.AdminRetrospectCountGetResponse;
 import org.layer.space.controller.dto.AdminSpaceCountGetResponse;
 import org.layer.space.controller.dto.AdminSpacesGetResponse;
 import org.layer.space.service.AdminSpaceService;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
 import java.time.LocalDateTime;
 
@@ -19,7 +17,7 @@ public class AdminSpaceController implements AdminSpaceApi {
 
     private final AdminSpaceService adminSpaceService;
     @Override
-    @GetMapping("/space")
+    @GetMapping("/")
     public ResponseEntity<AdminSpacesGetResponse> getSpaceData(
             @RequestParam("startDate") LocalDateTime startDate,
             @RequestParam("endDate") LocalDateTime endDate,
@@ -29,11 +27,21 @@ public class AdminSpaceController implements AdminSpaceApi {
     }
 
     @Override
-    @GetMapping("/space/count/user-only")
+    @GetMapping("/count/user-only")
     public ResponseEntity<AdminSpaceCountGetResponse> getSpaceCount(
             @RequestParam("startDate") LocalDateTime startDate,
             @RequestParam("endDate") LocalDateTime endDate,
             @RequestParam("password") String password) {
         return ResponseEntity.ok(adminSpaceService.getSpaceCount(startDate, endDate, password));
     }
+
+    @Override
+    @GetMapping("/{spaceId}/retrospect/count")
+    public ResponseEntity<AdminRetrospectCountGetResponse> getRetrospectCountInSpace(@RequestParam("startDate") LocalDateTime startDate,
+                                                                                     @RequestParam("endDate") LocalDateTime endDate,
+                                                                                     @PathVariable("spaceId") Long spaceId,
+                                                                                     @RequestParam("password") String password) {
+        return ResponseEntity.ok(adminSpaceService.getRetrospectCountInSpace(startDate, endDate, spaceId, password));
+    }
+
 }
